+++ CHANGELOG.rst
@@ -16,7 +16,9 @@ Added
     - ``RemoteSiteAppPlugin`` site plugin (#3)
     - ``PROJECTROLES_SITE_MODE`` setting (#3)
     - Remote site and project management site app (#3)
+    - Remote project API (#3)
     - Generic SODAR API base classes
+    - ``SodarUserMixin`` for SODAR user helpers in tests
 - **Timeline**
     - ``RemoteSite`` model in ``api.get_event_description()`` (#3)